Split classes to separate files
