+++ src/VirtoCommerce.ShipStationModule.Core/Models/ShipStationCustomer.cs
@@ -10,36 +10,3 @@ public class ShipStationCustomer
 
     public ShipStationShipTo ShipTo { get; set; }
 }
-
-public class ShipStationBillTo
-{
-    public XmlCDataSection Name { get; set; }
-
-    public XmlCDataSection Company { get; set; }
-
-    public XmlCDataSection Phone { get; set; }
-
-    public XmlCDataSection Email { get; set; }
-}
-
-public class ShipStationShipTo
-{
-    public XmlCDataSection Name { get; set; }
-
-    public XmlCDataSection Company { get; set; }
-
-    public XmlCDataSection Address1 { get; set; }
-
-    public XmlCDataSection Address2 { get; set; }
-
-    public XmlCDataSection City { get; set; }
-
-    public XmlCDataSection State { get; set; }
-
-    public XmlCDataSection PostalCode { get; set; }
-
-    public XmlCDataSection Country { get; set; }
-
-    public XmlCDataSection Phone { get; set; }
-
-}
